requirements.txt
pyopenssl
flask
rpi_hardware_pwm
spidev
RPi.GPIO
asyncio
netifaces
websockets
pyopenssl
cryptography